+++ CHANGES.md
@@ -17,6 +17,7 @@ profile. This started with version 0.26.0.
 - Fix extension-point spacing in structures (#2450, @Julow)
 - \* Consistent break after string constant argument (#2453, @Julow)
 - Fix invalid syntax generated with `ocp-indent-compat` (#2445, @Julow)
+- Fixed bug with attributes on sub-expressions of infix operators (#2459, @tdelvecchio-jsc)
 
 ## 0.26.1 (2023-09-15)